+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.16.0 (Unreleased)
+
+**Spotify Web API changes**
+- ([#556](https://github.com/ramsayleung/rspotify/pull/556)) Deprecate removed endpoints#556
+- ([#558](https://github.com/ramsayleung/rspotify/pull/558)) migrate deprecated endpoint to new endpoint
+
 ## 0.15.3 (2025.11.07)
 **Bugfixes**
 - ([#542](https://github.com/ramsayleung/rspotify/pull/542)) Fix: use unspecified ip